+++ ts/features/mvl/screens/details/components/attachment/MvlAttachments.tsx
@@ -2,19 +2,19 @@ import { View } from "native-base";
 import * as React from "react";
 import { StyleSheet, TouchableOpacity } from "react-native";
 import Svg from "react-native-svg";
-import Default from "../../../../../../img/features/mvl/attachmentsIcon/default.svg";
-import Pdf from "../../../../../../img/features/mvl/attachmentsIcon/pdf.svg";
-import { H2 } from "../../../../../components/core/typography/H2";
-import { H3 } from "../../../../../components/core/typography/H3";
-import { H5 } from "../../../../../components/core/typography/H5";
-import { IOColors } from "../../../../../components/core/variables/IOColors";
-import { IOStyles } from "../../../../../components/core/variables/IOStyles";
-import ItemSeparatorComponent from "../../../../../components/ItemSeparatorComponent";
-import IconFont from "../../../../../components/ui/IconFont";
-import I18n from "../../../../../i18n";
-import { formatByte } from "../../../../../types/digitalInformationUnit";
-import { MvlAttachment, MvlData } from "../../../types/mvlData";
-import { useDownloadAttachmentConfirmationBottomSheet } from "./attachment/DownloadAttachmentConfirmationBottomSheet";
+import Default from "../../../../../../../img/features/mvl/attachmentsIcon/default.svg";
+import Pdf from "../../../../../../../img/features/mvl/attachmentsIcon/pdf.svg";
+import { H2 } from "../../../../../../components/core/typography/H2";
+import { H3 } from "../../../../../../components/core/typography/H3";
+import { H5 } from "../../../../../../components/core/typography/H5";
+import { IOColors } from "../../../../../../components/core/variables/IOColors";
+import { IOStyles } from "../../../../../../components/core/variables/IOStyles";
+import ItemSeparatorComponent from "../../../../../../components/ItemSeparatorComponent";
+import IconFont from "../../../../../../components/ui/IconFont";
+import I18n from "../../../../../../i18n";
+import { formatByte } from "../../../../../../types/digitalInformationUnit";
+import { MvlAttachment, MvlData } from "../../../../types/mvlData";
+import { useDownloadAttachmentConfirmationBottomSheet } from "./DownloadAttachmentConfirmationBottomSheet";
 
 type Props = {
   attachments: MvlData["attachments"];
